pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbeats-tutorial"
version = "0.1.0"
description = "Agentbeats Tutorial"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.5",
    "audioop-lts>=0.2.2",
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "fastmcp>=2.14.0",
    "google-adk>=1.14.1",
    "google-genai>=1.36.0",
    "gymnasium>=1.2.3",
    "ipython>=9.8.0",
    "jupyter-core>=5.9.1",
    "litellm>=1.0.0",
    "loguru>=0.7.0",
    "lxml>=6.0.2",
    "mammoth>=1.11.0",
    "markdownify>=1.2.2",
    "mcp[cli]>=1.0.0",
    "nbconvert>=7.16.6",
    "notebook>=7.5.1",
    "openpyxl>=3.1.5",
    "pdfminer>=20191125",
    "pdfminer-six>=20251230",
    "puremagic>=1.30",
    "pydantic>=2.11.9",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "speechrecognition>=3.14.5",
    "tau2 @ git+https://github.com/sierra-research/tau2-bench.git",
    "tiktoken>=0.12.0",
    "uvicorn>=0.35.0",
    "youtube-transcript-api>=1.2.3",
]

[project.scripts]
agentbeats-run = "agentbeats.run_scenario:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.18.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agentbeats"]
